Moved game-field component to app, moved game loop logic to game-field
