+++ .vscode/extensions.json
@@ -2,6 +2,6 @@
   "recommendations": [
     "nrwl.angular-console",
     "esbenp.prettier-vscode",
-    "dbaeumer.vscode-eslint"
+    "gruntfuggly.todo-tree",
   ]
 }
